adds `retain_s3_files` and `s3_file_naming_scheme`

+++ target_snowflake/__init__.py
@@ -359,6 +359,7 @@ def flush_streams(
             row_count=row_count,
             db_sync=stream_to_sync[stream],
             no_compression=config.get('no_compression'),
+            retain_s3_files=config.get('retain_s3_files'),
             delete_rows=config.get('hard_delete'),
             temp_dir=config.get('temp_dir')
         ) for stream in streams_to_flush)
@@ -385,11 +386,11 @@ def flush_streams(
     return flushed_state
 
 
-def load_stream_batch(stream, records_to_load, row_count, db_sync, no_compression=False, delete_rows=False,
-                      temp_dir=None):
+def load_stream_batch(stream, records_to_load, row_count, db_sync, no_compression=False,
+                      retain_s3_files=False, delete_rows=False, temp_dir=None):
     # Load into snowflake
     if row_count[stream] > 0:
-        flush_records(stream, records_to_load, row_count[stream], db_sync, temp_dir, no_compression)
+        flush_records(stream, records_to_load, row_count[stream], db_sync, temp_dir, no_compression, retain_s3_files)
 
         # Delete soft-deleted, flagged rows - where _sdc_deleted at is not null
         if delete_rows:
@@ -405,17 +406,19 @@ def write_record_to_file(outfile, records_to_load, record_to_csv_line_transforme
         outfile.write(bytes(csv_line + '\n', 'UTF-8'))
 
 
-def flush_records(stream, records_to_load, row_count, db_sync, temp_dir=None, no_compression=False):
+def flush_records(stream, records_to_load, row_count, db_sync, temp_dir=None,
+                  no_compression=False, retain_s3_files=False):
     if temp_dir:
         os.makedirs(temp_dir, exist_ok=True)
-    csv_fd, csv_file = mkstemp(suffix='.csv', prefix='records_', dir=temp_dir)
     record_to_csv_line_transformer = db_sync.record_to_csv_line
 
     # Using gzip or plain file object
     if no_compression:
+        csv_fd, csv_file = mkstemp(suffix='.csv', prefix='records_', dir=temp_dir)
         with open(csv_fd, 'wb') as outfile:
             write_record_to_file(outfile, records_to_load, record_to_csv_line_transformer)
     else:
+        csv_fd, csv_file = mkstemp(suffix='.csv.gz', prefix='records_', dir=temp_dir)
         with gzip.open(csv_file, 'wb') as outfile:
             write_record_to_file(outfile, records_to_load, record_to_csv_line_transformer)
 
@@ -424,7 +427,8 @@ def flush_records(stream, records_to_load, row_count, db_sync, temp_dir=None, no
     db_sync.load_csv(s3_key, row_count, size_bytes)
 
     os.remove(csv_file)
-    db_sync.delete_from_stage(s3_key)
+    if not retain_s3_files:
+        db_sync.delete_from_stage(s3_key)
 
 
 def main():
